# patients-portal-api/Controllers/VisitsApi.cs
/*
 * Patients Portal
 *
 * Best Medical Portal API ever
 *
 * OpenAPI spec version: 1.0.0
 * Contact: nikolas.tsk@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using eu.fiit.PatientsPortal.Attributes;

using Microsoft.AspNetCore.Authorization;
using eu.fiit.PatientsPortal.Models;

namespace eu.fiit.PatientsPortal.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class VisitsApiController : ControllerBase
    { 
        /// <summary>
        /// Add a new Visit
        /// </summary>
        /// <param name="body">Visit object</param>
        /// <response code="405">Invalid input</response>
        [HttpPost]
        [Route("/yimejky/PatientsPortal/1.0.0/visits")]
        [ValidateModelState]
        [SwaggerOperation("AddVisit")]
        public virtual IActionResult AddVisit([FromBody]Visit body)
        { 
            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Deletes a Visit
        /// </summary>
        /// <param name="visitId">Visit id to delete</param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Visit not found</response>
        [HttpDelete]
        [Route("/yimejky/PatientsPortal/1.0.0/visits/{visitId}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteVisit")]
        public virtual IActionResult DeleteVisit([FromRoute][Required]long? visitId)
        { 
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get all visits
        /// </summary>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("/yimejky/PatientsPortal/1.0.0/visits")]
        [ValidateModelState]
        [SwaggerOperation("GetVisits")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Visit>), description: "successful operation")]
        public virtual IActionResult GetVisits()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Visit>));
            string exampleJson = null;
            exampleJson = "[ {\n  \"id\" : \"id\"\n}, {\n  \"id\" : \"id\"\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Visit>>(exampleJson)
                        : default(List<Visit>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update an existing Visit
        /// </summary>
        /// <param name="body">Visit object</param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Visit not found</response>
        [HttpPut]
        [Route("/yimejky/PatientsPortal/1.0.0/visits")]
        [ValidateModelState]
        [SwaggerOperation("UpdateVisit")]
        public virtual IActionResult UpdateVisit([FromBody]Visit body)
        { 
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);

            throw new NotImplementedException();
        }
    }
}